producers/erp/requirements.txt
kafka-python
pymysql
cryptography